SPARK-38617 rename SqlCompilerResourceSuite and save String names for REST/UI Output

+++ sql/core/src/main/scala/org/apache/spark/status/api/v1/sql/api.scala
@@ -41,7 +41,7 @@ class CompileData private[spark] (
 
 case class Rule private[spark](
     ruleName: String,
-    timeMs: Long,
+    timeMs: Double,
     numInvocations: Long,
     numEffectiveInvocations: Long)
